Add + enable archive job
Also updated some to-dos:
- ticked off "write archive job"
- removed the deduplication todo since `zstd` does it well enough

And added the example crontab entry for running the archiver job
every Sunday morning at 3am local time.  (That's derived from my
personal schedule since I am very unlikely to be awake and using
the server at that point; this should be different if the server
admin works a night shift or a Sun-Fri schedule or whatever.